Omo e dey work o

+++ internal/models/todo.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Todo struct {
-	ID        int64
+	ID        int
 	Name      string
 	Details   string
 	Completed bool
@@ -18,40 +18,32 @@ type Queries struct {
 	DB *sql.DB
 }
 
-const createTodo = `-- name: CreateTodo :execresult
-INSERT INTO todo (
-  name, details, completed
-) VALUES (
-  ?, ?, ?
-)
-`
+const createTodo = `INSERT INTO todo (name, details, completed, user_id) 
+VALUES (?, ?, ?, ?)`
 
 type CreateTodoParams struct {
 	Name      string
 	Details   string
 	Completed bool
+	UserID    int
 }
 
 func (q *Queries) CreateTodo(ctx context.Context, arg CreateTodoParams) (sql.Result, error) {
-	return q.DB.ExecContext(ctx, createTodo, arg.Name, arg.Details, arg.Completed)
+	return q.DB.ExecContext(ctx, createTodo, arg.Name, arg.Details, arg.Completed, arg.UserID)
 }
 
-const deleteTodo = `-- name: DeleteTodo :exec
-DELETE FROM todo WHERE id = ?
-`
+const deleteTodo = `DELETE FROM todo WHERE id = ? AND user_id = ?`
 
-func (q *Queries) DeleteTodo(ctx context.Context, id int64) error {
-	_, err := q.DB.ExecContext(ctx, deleteTodo, id)
+func (q *Queries) DeleteTodo(ctx context.Context, id, user_id int) error {
+	_, err := q.DB.ExecContext(ctx, deleteTodo, id, user_id)
 	return err
 }
 
-const getTodo = `-- name: GetTodo :one
-SELECT id, name, details, completed, created FROM todo
-WHERE id = ? LIMIT 1
-`
+const getTodo = `SELECT id, name, details, completed, created FROM todo
+WHERE id = ? AND user_id = ? LIMIT 1`
 
-func (q *Queries) GetTodo(ctx context.Context, id int64) (Todo, error) {
-	row := q.DB.QueryRowContext(ctx, getTodo, id)
+func (q *Queries) GetTodo(ctx context.Context, id, user_id int) (Todo, error) {
+	row := q.DB.QueryRowContext(ctx, getTodo, id, user_id)
 	var i Todo
 	err := row.Scan(
 		&i.ID,
@@ -63,14 +55,12 @@ func (q *Queries) GetTodo(ctx context.Context, id int64) (Todo, error) {
 	return i, err
 }
 
-const listTodo = `-- name: ListTodo :many
-SELECT id, name, details, completed, created FROM todo
-WHERE name != ''
-ORDER BY created
-`
+const listTodo = `SELECT id, name, details, completed, created FROM todo
+WHERE name != '' AND user_id = ?
+ORDER BY created`
 
-func (q *Queries) ListTodo(ctx context.Context) ([]Todo, error) {
-	rows, err := q.DB.QueryContext(ctx, listTodo)
+func (q *Queries) ListTodo(ctx context.Context, userID int) ([]Todo, error) {
+	rows, err := q.DB.QueryContext(ctx, listTodo, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -98,16 +88,14 @@ func (q *Queries) ListTodo(ctx context.Context) ([]Todo, error) {
 	return items, nil
 }
 
-const updateTodo = `-- name: UpdateTodo :exec
-UPDATE todo SET name = ?, details = ?, completed = ? 
-WHERE id = ?
-`
+const updateTodo = `UPDATE todo SET name = ?, details = ?, completed = ? WHERE id = ? AND user_id = ?`
 
 type UpdateTodoParams struct {
 	Name      string
 	Details   string
 	Completed bool
-	ID        int64
+	ID        int
+	UserID    int
 }
 
 func (q *Queries) UpdateTodo(ctx context.Context, arg UpdateTodoParams) error {
@@ -116,6 +104,7 @@ func (q *Queries) UpdateTodo(ctx context.Context, arg UpdateTodoParams) error {
 		arg.Details,
 		arg.Completed,
 		arg.ID,
+		arg.UserID,
 	)
 	return err
 }
